Uses grc-web.StatusCode instead of grpc.status (#401)
The SDK doesn't work in the wallet right now because the web version is pulling in `grpc`, which is `node` only.

I did a bisect and ended up at https://github.com/xpring-eng/Xpring-JS/pull/328.

As such, I replaced all uses of `grpc.status` with `grpcStatus`. The values for it come from https://github.com/grpc/grpc/blob/master/doc/statuscodes.md - shouts out to `UNAUTHENTICATED` for being interleaved instead of where it would go monotonically.

Also note that `ilp-error` still includes `ServiceError`, which might seem kind of weird. It is ONLY being used as a type, which is compiled away, which is why it works. We might want to consider making a type just for `ServiceError`, since it's just a bag of values.

I tested this by running tests and building the wallet against my `xpring-js` copy locally.